refactor(game-session): finish ActiveGameSession split
Moves journal, ending, and choice handlers into dedicated
hooks so the main component stays under the 300-line
guideline while keeping behavior the same.

Closes #989
